+++ pages/api/gallery/create.js
@@ -1,0 +1,37 @@
+import auth from "../../../src/server/services/authorization-service";
+import dbConnect from "../../../utils/dbConnect";
+import Gallery from "../../../models/galleryModel";
+import User from "../../../models/userModel";
+
+export default async function handler(req, res) {
+  const { method } = req;
+  await dbConnect();
+
+  switch (method) {
+    case "POST":
+      try {
+        const jwt = auth(req.headers.authorization, res, (jwt) => jwt);
+        const user = await User.findById(jwt.id);
+        if (!user) return res.json({ error: "Authentication error" });
+
+        const gallery = new Gallery({
+          user: jwt.id,
+          name: req.body.name,
+          description: req.body.description,
+          nfts: [],
+        });
+
+        gallery.save();
+        return res.json({ success: gallery });
+      } catch (error) {
+        console.error(error);
+        res.status(400).json({ success: false, message: "fetch failed" });
+      }
+
+      break;
+
+    default:
+      res.status(400).json({ success: false, message: "route not matched" });
+      break;
+  }
+}
