+++ pages/api/gallery/edit.js
@@ -1,3 +1,4 @@
+import auth from "../../../src/server/services/authorization-service";
 import dbConnect from "../../../utils/dbConnect";
 import Gallery from "../../../models/galleryModel";
 import User from "../../../models/userModel";
@@ -13,21 +14,15 @@ export default async function handler(req, res) {
         const user = await User.findById(jwt.id);
         if (!user) return res.json({ error: "Authentication error" });
 
-        const galleries = await Gallery.find({ user: jwt.id });
+        const gallery = await Gallery.findById(req.body.gallery);
+        if (!gallery) return res.json({ error: "Could not find gallery" });
+        if (!gallery.user.equals(jwt.id))
+          return res.json({ error: "Authentication error" });
 
-        if (!req.body.galleries || !req.body.galleries.length)
-          return res.json({ error: "Issue reordering" });
-
-        req.body.galleries.forEach((gallery) => {
-          galleries.forEach((find) => {
-            if (find.id === gallery.id) {
-              find.order = gallery.order;
-              find.save();
-            }
-          });
-        });
-
-        return res.json({ success: "Gallery order updated" });
+        gallery.name = req.body.name;
+        gallery.description = req.body.description;
+        gallery.save();
+        return res.json({ success: "Gallery updated" });
       } catch (error) {
         console.error(error);
         res.status(400).json({ success: false, message: "fetch failed" });
