+++ pages/api/gallery/remove.js
@@ -1,3 +1,4 @@
+import auth from "../../../src/server/services/authorization-service";
 import dbConnect from "../../../utils/dbConnect";
 import Gallery from "../../../models/galleryModel";
 import User from "../../../models/userModel";
@@ -18,18 +19,13 @@ export default async function handler(req, res) {
         if (!gallery.user.equals(jwt.id))
           return res.json({ error: "Authentication error" });
 
-        if (!req.body.nfts || !req.body.nfts.length)
-          return res.json({ error: "Issue reordering" });
+        const index = gallery.nfts.findIndex((e) => e.id === req.body.id);
+        if (index >= 0) {
+          gallery.nfts.splice(index, 1);
+          gallery.save();
+        }
 
-        req.body.nfts.forEach((nft) => {
-          gallery.nfts.forEach((find) => {
-            if (find.id === nft.id) find.order = nft.order;
-          });
-        });
-
-        gallery.save();
-
-        return res.json({ success: "NFT order updated" });
+        return res.json({ success: "NFT removed" });
       } catch (error) {
         console.error(error);
         res.status(400).json({ success: false, message: "fetch failed" });
