+++ pages/api/gallery/renderGalleries.js
@@ -1,3 +1,4 @@
+import auth from "../../../src/server/services/authorization-service";
 import dbConnect from "../../../utils/dbConnect";
 import Gallery from "../../../models/galleryModel";
 import User from "../../../models/userModel";
@@ -13,8 +14,21 @@ export default async function handler(req, res) {
         const user = await User.findById(jwt.id);
         if (!user) return res.json({ error: "Authentication error" });
 
-        await Gallery.findOneAndRemove({ _id: req.body.gallery, user: jwt.id });
-        return res.json({ success: "Gallery deleted" });
+        const galleries = await Gallery.find({ user: jwt.id });
+
+        if (!req.body.galleries || !req.body.galleries.length)
+          return res.json({ error: "Issue reordering" });
+
+        req.body.galleries.forEach((gallery) => {
+          galleries.forEach((find) => {
+            if (find.id === gallery.id) {
+              find.order = gallery.order;
+              find.save();
+            }
+          });
+        });
+
+        return res.json({ success: "Gallery order updated" });
       } catch (error) {
         console.error(error);
         res.status(400).json({ success: false, message: "fetch failed" });
