+++ pages/api/gallery/reorderCollection.js
@@ -1,3 +1,4 @@
+import auth from "../../../src/server/services/authorization-service";
 import dbConnect from "../../../utils/dbConnect";
 import Gallery from "../../../models/galleryModel";
 import User from "../../../models/userModel";
@@ -18,13 +19,18 @@ export default async function handler(req, res) {
         if (!gallery.user.equals(jwt.id))
           return res.json({ error: "Authentication error" });
 
-        const index = gallery.nfts.findIndex((e) => e.id === req.body.id);
-        if (index >= 0) {
-          gallery.nfts.splice(index, 1);
-          gallery.save();
-        }
+        if (!req.body.nfts || !req.body.nfts.length)
+          return res.json({ error: "Issue reordering" });
 
-        return res.json({ success: "NFT removed" });
+        req.body.nfts.forEach((nft) => {
+          gallery.nfts.forEach((find) => {
+            if (find.id === nft.id) find.order = nft.order;
+          });
+        });
+
+        gallery.save();
+
+        return res.json({ success: "NFT order updated" });
       } catch (error) {
         console.error(error);
         res.status(400).json({ success: false, message: "fetch failed" });
